increase MAX_IF to 100
Several users of OPNsense and pfSense report a persistent error
with their configuration: "There must be at least 1 Vif as upstream."

It seems that there is no real error in igmpproxy, but rather with
the way that network addresses are looked up prior to starting up.
This can break when many addresses are present in the system.

There is no exact science for the increase to 100 and previous custom
patches on both projects have had this raised to 80 back when version
0.1 was used.  There could be further work required to dynamically
parse addresses in a way that this raised limit does not cause further
issues in the future.

See also: https://github.com/opnsense/plugins/issues/590
See also: https://redmine.pfsense.org/issues/8935